Separate entities and schemas
