+++ src/main.ts
@@ -3,17 +3,24 @@ import { startServer } from './server'
 import { createLogger } from './logger'
 import { close, connect } from './database'
 import { initSubmissions } from './submission/mongoRepository'
-import { initRegistry, loadRegistry } from './registry'
+import { initSchemas, loadSchemas } from './schema/mongoRepository'
+import { initEntities, loadEntities } from './entity/mongoRepository'
 
 const logger = createLogger(__filename)
 
 async function main() {
   logger.info(`Starting app with the following config`, hideSecrets(config))
 
+  // TODO
+  // Automatically add audit attributes and validate entity
+  // Automatically add audit attributes and validate schema
+
   const database = await connect(config.db)
   initSubmissions(database)
-  initRegistry(database)
-  await loadRegistry()
+  initSchemas(database)
+  initEntities(database)
+  await loadSchemas()
+  await loadEntities()
   startServer(config.server)
   process.on('SIGINT', shutdownGracefully())
   process.on('SIGTERM', shutdownGracefully())
